Source/app/config.py
from os import getenv
import dotenv

# Load secrets from environment
dotenv.load_dotenv()

# Load Telegram bot API token
TELEGRAM_TOKEN = getenv("TELEGRAM_TOKEN")
# Pyrogram session string
SESSION_STRING = getenv("SESSION_STRING")
# Load public domain for webhooks
WEBHOOK_DOMAIN = getenv("WEBHOOK_DOMAIN")
